[Refactor] Fix failing /Collect2 tests
